--- src/Components/ContainerHeader.jsx
@@ -1,11 +1,0 @@
-import ViewHeader from "./ViewHeader"
-
-const ContainerHeader = () => {
-  return (
-    <div>
-      <ViewHeader />
-    </div>
-  )
-}
-
-export default ContainerHeader
