--- src/Components/ContainerShiftPanel.jsx
@@ -1,11 +1,0 @@
-import ViewShiftPanel from "./ViewShiftPanel"
-
-const ContainerShiftPanel = () => {
-  return (
-    <div>
-      <ViewShiftPanel />
-    </div>
-  )
-}
-
-export default ContainerShiftPanel
